Create Breakfast Cereal Analysis
Customer Rating of Breakfast Cereals. The dataset Cereals.csv includes nutritional
information, store display, and consumer ratings for 77 breakfast cereals.
Data Preprocessing. Remove all cereals with missing values.
a. Apply hierarchical clustering to the data using Euclidean distance to the normalized
measurements. Compare the dendrograms from single linkage and complete
linkage, and look at cluster centroids. Comment on the structure of the clusters
and on their stability. Hint: To obtain cluster centroids for hierarchical clustering,
compute the average values of each cluster members, using the aggregate() function.
b. Which method leads to the most insightful or meaningful clusters?
c. Choose one of the methods. How many clusters would you use? What distance is
used for this cutoff? (Look at the dendrogram.)
d. The elementary public schools would like to choose a set of cereals to include in
their daily cafeterias. Every day a different cereal is offered, but all cereals should
support a healthy diet. For this goal, you are requested to find a cluster of “healthy
cereals.” Should the data be normalized? If not, how should they be used in the
cluster analysis?